move channel-permission-profiles from config/ to channels/
